Rename `alternative_stack` to `fault`.
