+++ lib/runtime-core/image-loading-linux-x86-64.s
@@ -1,4 +1,4 @@
-# NOTE: Keep this consistent with `alternative_stack.rs`.
+# NOTE: Keep this consistent with `fault.rs`.
 
 .globl run_on_alternative_stack
 run_on_alternative_stack:
