+++ lib/runtime-core/image-loading-macos-x86-64.s
@@ -1,4 +1,4 @@
-# NOTE: Keep this consistent with `alternative_stack.rs`.
+# NOTE: Keep this consistent with `fault.rs`.
 
 .globl _run_on_alternative_stack
 _run_on_alternative_stack:
