+++ lib/runtime-core/src/state.rs
@@ -362,7 +362,7 @@ impl InstanceImage {
 #[cfg(all(unix, target_arch = "x86_64"))]
 pub mod x64 {
     use super::*;
-    use crate::alternative_stack::{catch_unsafe_unwind, run_on_alternative_stack};
+    use crate::fault::{catch_unsafe_unwind, run_on_alternative_stack};
     use crate::codegen::BkptMap;
     use crate::structures::TypedIndex;
     use crate::types::LocalGlobalIndex;
