+++ lib/singlepass-backend/src/protect_unix.rs
@@ -11,7 +11,7 @@
 //!
 use std::any::Any;
 use std::cell::Cell;
-use wasmer_runtime_core::alternative_stack::{
+use wasmer_runtime_core::fault::{
     begin_unsafe_unwind, catch_unsafe_unwind, ensure_sighandler,
 };
 use wasmer_runtime_core::codegen::BkptMap;
