+++ src/bin/wasmer.rs
@@ -514,7 +514,7 @@ fn execute_wasm(options: &Run) -> Result<(), String> {
             #[cfg(feature = "backend:singlepass")]
             unsafe {
                 if options.backend == Backend::Singlepass {
-                    use wasmer_runtime_core::alternative_stack::{
+                    use wasmer_runtime_core::fault::{
                         catch_unsafe_unwind, ensure_sighandler,
                     };
                     use wasmer_runtime_core::state::{
